refactor: split medical_equipment_service.py (1033 LOC → 4 modules)
Split into app/services/medical_equipment/ package:
- _base.py: imports, enums (DeviceType, DeviceStatus, ConnectionType), mixin base
- _core.py: 24 methods (connect, disconnect, __init__, drivers)
- _operations.py: 24 methods (measurements, calibration, queries)
- __init__.py: composes CoreMixin + OperationsMixin

Shim re-exports: MedicalEquipmentService, ConnectionType, DeviceStatus,
DeviceType, get_medical_equipment_service.

Added F811 to ruff ignore (intentional in mixin pattern — methods
defined in multiple mixins).

All 34 tests pass. Ruff: 0 errors. API router: 1128 routes, 0 duplicates.
